Automates Docker image build and push process
Introduces automated Docker image building and pushing to a registry.

Adds `.env.example` and `.gitignore` files for configuration and exclusion.

Updates the `build_and_push.yml` workflow to handle dynamic image tags and registry authentication.

The deploy script is modified to utilize the new environment variables and the tag from the .tag file.

Dockerfile label is updated to reflect compatibility with Deployer.org